Remove set -euo pipefail
Although I'd prefer to keep this it leads to hard exit codes that can't
be surpressed for example when the plugin is called on a buffer where no
url is present. Since the script mostly operates on variables it should
be safe to omit these options without risking any major issues.